+++ projects/ontimize-web-ngx/src/lib/components/app-header/o-app-header.component.ts
@@ -1,4 +1,5 @@
 import { Component, ElementRef, EventEmitter, Injector, OnDestroy, ViewEncapsulation } from '@angular/core';
+import { ThemePalette } from '@angular/material';
 import { Router } from '@angular/router';
 import { Subscription } from 'rxjs';
 
@@ -10,7 +11,8 @@ import { ServiceUtils } from '../../util/service.utils';
 export const DEFAULT_INPUTS_O_APP_HEADER = [
   'showUserInfo: show-user-info',
   'showLanguageSelector: show-language-selector',
-  'useFlagIcons: use-flag-icons'
+  'useFlagIcons: use-flag-icons',
+  'color'
 ];
 
 export const DEFAULT_OUTPUTS_O_APP_HEADER = [
@@ -44,6 +46,8 @@ export class OAppHeaderComponent implements OnDestroy {
   useFlagIcons: boolean = false;
 
   public onSidenavToggle = new EventEmitter<void>();
+
+  private _color: ThemePalette;
 
   constructor(
     protected router: Router,
@@ -80,6 +84,14 @@ export class OAppHeaderComponent implements OnDestroy {
 
   get showHeaderTitle(): boolean {
     return this._headerTitle.length > 0;
+  }
+
+  set color(newValue: ThemePalette) {
+    this._color = newValue;
+  }
+
+  get color(): ThemePalette {
+    return this._color;
   }
 
 }